stend/requirements.txt
fastapi
uvicorn
requests
websocket-client
pydantic
python-multipart
